Fix bug in Trapezoid3D method.
It seems like the factor "oldSum" has been wrongly added when the Integral method was translated from old MMC java code to the new PROPOSAL C++ code.
When randomNumber_ is positive, we expect randomNumber_ to be a relative value.
When randomNumber_ is negative, we expect randomNumber_ to be an absolute value, hence the factor oldSum is not necessary in this case.